WIP: Safestrap v3 changes

+++ gui/action_safestrap.cpp
@@ -8,32 +8,51 @@ int createImagePartition(string slotName, string imageName, int imageSize, strin
 	sprintf(cmd, "/%s", mountName.c_str());
 	ensure_path_unmounted(cmd);
 	sprintf(cmd, "rm -rf /ss/safestrap/%s/%s.img", slotName.c_str(), imageName.c_str());
-	__system(cmd);
+	if (__system(cmd) != 0) return 1;
+
 	sprintf(cmd, "rm /dev/block/%s", imageName.c_str());
-	__system(cmd);
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+
 	sprintf(cmd, "/sbin/bbx losetup -d /dev/block/loop%d", loopNum);
-	__system(cmd);
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
 	DataManager::SetValue("ui_progress", progressBase1);
 
 	DataManager::SetValue("tw_operation", "Creating " + imageName + ".img...");
 	ui_print("Creating %s.img...\n", imageName.c_str());
 
 	sprintf(cmd, "dd if=/dev/zero of=/ss/safestrap/%s/%s.img bs=1M count=%d", slotName.c_str(), imageName.c_str(), imageSize);
-	__system(cmd);
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+	sprintf(cmd, "/sbin/fsync /ss/safestrap/%s/%s.img", slotName.c_str(), imageName.c_str());
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
 	DataManager::SetValue("ui_progress", progressBase2);
 
 	DataManager::SetValue("tw_operation", "Writing ext3 filesystem on " + imageName + "...");
 	ui_print("Writing ext3 filesystem on %s...\n", imageName.c_str());
 	sprintf(cmd, "/sbin/bbx losetup /dev/block/loop%d /ss/safestrap/%s/%s.img", loopNum, slotName.c_str(), imageName.c_str());
-	__system(cmd);
 	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+
 	sprintf(cmd, "ln -s /dev/block/loop%d /dev/block/%s", loopNum, imageName.c_str());
-	__system(cmd);
-	sprintf(cmd, "mkfs.ext2 /dev/block/%s", imageName.c_str());
-	__system(cmd);
 	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+
+	sprintf(cmd, "mkfs.ext2 /dev/block/%s", imageName.c_str());
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+	sprintf(cmd, "/sbin/fsync /ss/safestrap/%s/%s.img", slotName.c_str(), imageName.c_str());
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+
 	sprintf(cmd, "tune2fs -j /dev/block/%s", imageName.c_str());
-	__system(cmd);
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
+	sprintf(cmd, "/sbin/fsync /ss/safestrap/%s/%s.img", slotName.c_str(), imageName.c_str());
+	usleep(100000);
+	if (__system(cmd) != 0) return 1;
 	DataManager::SetValue("ui_progress", progressBase3);
 
 	return 0;
@@ -46,6 +65,15 @@ int GUIAction::doSafestrapAction(Action action, int isThreaded /* = 0 */) {
 	DataManager::GetValue(TW_SIMULATE_ACTIONS, simulate);
 
 	LOGD("doSafestrapAction(%s(%s, %d));\n", function.c_str(), arg.c_str(), isThreaded);
+
+	if (function == "refreshsizesnt")
+	{
+		if (simulate) {
+			simulate_progress_bar();
+		} else
+			update_system_details();
+		return 0;
+	}
 
 	if (function == "changeslot") {
 		char cmd[512];
@@ -78,15 +106,106 @@ int GUIAction::doSafestrapAction(Action action, int isThreaded /* = 0 */) {
 		if (lSize > 0) {
 			fseek(in, 0, SEEK_SET);
 			// adjust for EOF
-			lSize -= 1;
 			result = fread(array, 1, lSize, in);
 			LOGD("result=%d\n",(long)result);
-			if ((long)result == lSize)
+			if ((long)result == lSize) {
+				array[lSize-1] = '\0';
 				str = array;
+			}
 		}
 		fclose(in);
 //		ui_print("bootslot set: %s\n", str.c_str());
 		DataManager::SetValue(arg, str);
+		return 0;
+	}
+
+	if (function == "checkslotnames") {
+		char array[512];
+		string str = "Undefined";
+		string var = "";
+
+		for (int i = 1; i < 5; i++) {
+			sprintf(array, "rom-slot%d", i);
+			var = array;
+			DataManager::GetValue("tw_" + var + "_name", str);
+			fprintf(stderr, "checkslotnames: tw_%s_name=%s\n", var.c_str(), str.c_str());
+		}
+		return 0;	
+	}
+
+	if (function == "readslotnames") {
+		int i;
+		char array[512];
+		char filename[512];
+		long lSize;
+		size_t result;
+		string str = "Undefined";
+		string var = "";
+
+		DataManager::SetValue("tw_rom-slot1_name", "ROM Slot 1");
+		DataManager::SetValue("tw_rom-slot2_name", "ROM Slot 2");
+		DataManager::SetValue("tw_rom-slot3_name", "ROM Slot 3");
+		DataManager::SetValue("tw_rom-slot4_name", "ROM Slot 4");
+
+		for (int i = 1; i < 5; i++) {
+			sprintf(array, "rom-slot%d", i);
+			var = array;
+
+			// Read in the file, if possible
+			sprintf(filename, "/ss/safestrap/rom-slot%d/name", i);
+			FILE* in = fopen(filename, "rb");
+			if (in) {
+				// obtain file size:
+				fseek(in, 0, SEEK_END);
+				lSize = ftell(in);
+				LOGD("lSize=%d\n", lSize);
+				if (lSize > 0) {
+					fseek(in, 0, SEEK_SET);
+					// adjust for EOF
+					result = fread(array, 1, lSize, in);
+					LOGD("result=%d\n",(long)result);
+					if ((long)result == lSize) {
+						array[lSize-1] = '\0';
+						str = array;
+					}
+				}
+				fclose(in);
+				fprintf(stderr, "found: tw_%s_name=%s\n", var.c_str(), str.c_str());
+				DataManager::SetValue("tw_" + var + "_name", str);
+			}
+			else {
+				DataManager::GetValue("tw_" + var + "_name", str);
+				fprintf(stderr, "not found: tw_%s_name=%s\n", var.c_str(), str.c_str());
+			}
+		}
+		return 0;
+	}
+
+	if (function == "setslotvarname") {
+		char array[512];
+		string str = arg;
+		string var = "tw_" + arg +"_name";
+		DataManager::GetValue(var, str);
+		fprintf(stderr, "Loaded from %s, setting tw_slotname: %s\n", var.c_str(), str.c_str());
+		DataManager::SetValue("tw_slotname", str);
+		return 0;
+	}
+
+	if (function == "setslotnickname") {
+		char cmd[512];
+		string romslot;
+		string slotname;
+
+		DataManager::GetValue("tw_trybootslot", romslot);
+		DataManager::GetValue("tw_try_slotname", slotname);
+		if ((romslot == "stock") || (slotname == ""))
+			return 0;
+
+		fprintf(stderr, "setslotnickname: echo \"%s\" > /ss/safestrap/%s/name\n", slotname.c_str(), romslot.c_str());
+		sprintf(cmd, "echo \"%s\" > /ss/safestrap/%s/name", slotname.c_str(), romslot.c_str());
+		__system(cmd);
+
+		DataManager::SetValue("tw_" + romslot + "_name", slotname);
 		return 0;
 	}
 
@@ -101,29 +220,6 @@ int GUIAction::doSafestrapThreadedAction(Action action, int isThreaded /* = 0 */
 	DataManager::GetValue(TW_SIMULATE_ACTIONS, simulate);
 	DataManager::GetValue("tw_trybootslot", arg);
 
-	if (function == "nandroid") {
-		operation_start("Nandroid");
-
-		if (simulate) {
-			DataManager::SetValue("tw_partition", "Simulation");
-			simulate_progress_bar();
-		}
-		else {
-			if (arg == "backup") {
-				nandroid_back_exe();
-				DataManager::SetValue(TW_BACKUP_NAME, "(Current Date)");
-			}
-			else if (arg == "restore")
-				nandroid_rest_exe();
-			else {
-				operation_end(1, simulate);
-				return -1;
-			}
-		}
-		operation_end(0, simulate);
-		return 0;
-	}
-
 	if (function == "createslot") {
 		operation_start("createslot");
 		if (arg == "stock")
@@ -135,7 +231,7 @@ int GUIAction::doSafestrapThreadedAction(Action action, int isThreaded /* = 0 */
 		}
 		else {
 			int system_size = 500;
-			int data_size = 1024;
+			int data_size = 500;
 			int cache_size = 500;
 			char cmd[512];
 
@@ -151,16 +247,44 @@ int GUIAction::doSafestrapThreadedAction(Action action, int isThreaded /* = 0 */
 			__system(cmd);
 
 			// SYSTEM
-			createImagePartition(arg, "system", system_size, "system", 7, 5, 20, 30);
+			if (createImagePartition(arg, "system", system_size, "system", 7, 5, 20, 30) != 0) {
+				DataManager::SetValue("tw_operation", "Error creating system partition!");
+				ui_print("Error creating system partition!\n");
+				ui_print("Cleaning up files...\n");
+				sprintf(cmd, "rm -rf /ss/safestrap/%s", arg.c_str());
+				__system(cmd);
+				return -1;
+			}
 
 			// USERDATA
-			createImagePartition(arg, "userdata", data_size, "data", 6, 35, 60, 70);
+			if (createImagePartition(arg, "userdata", data_size, "data", 6, 35, 60, 70) != 0) {
+				DataManager::SetValue("tw_operation", "Error creating data partition!");
+				ui_print("Error creating data partition!\n");
+				ui_print("Cleaning up files...\n");
+				sprintf(cmd, "rm -rf /ss/safestrap/%s", arg.c_str());
+				__system(cmd);
+				return -1;
+			}
 
 			// CACHE
 			ensure_path_mounted("/cache");
 			__system("cp -R /cache/recovery /tmp");
 			__system("rm -rf /cache/recovery");
-			createImagePartition(arg, "cache", cache_size, "cache", 5, 75, 85, 90);
+			ensure_path_unmounted("/cache");
+
+			if (createImagePartition(arg, "cache", cache_size, "cache", 5, 75, 85, 90) != 0) {
+				DataManager::SetValue("tw_operation", "Error creating cache partition!");
+				ui_print("Error creating cache partition!\n");
+				ui_print("Cleaning up files...\n");
+				sprintf(cmd, "rm -rf /ss/safestrap/%s", arg.c_str());
+				__system(cmd);
+				ensure_path_mounted("/cache");
+				__system("rm -rf /cache/recovery");
+				__system("cp -R /tmp/recovery /cache");
+				__system("rm -rf /tmp/recovery");
+				return -1;
+			}
+
 			ensure_path_mounted("/cache");
 			__system("rm -rf /cache/recovery");
 			__system("cp -R /tmp/recovery /cache");
@@ -185,5 +309,23 @@ int GUIAction::doSafestrapThreadedAction(Action action, int isThreaded /* = 0 */
 		return 0;
 	}
 
+	if (function == "deleteslot") {
+		char cmd[512];
+
+		operation_start("deleteslot");
+		if (arg == "stock")
+			return 0;
+
+		if (simulate) {
+			simulate_progress_bar();
+			ui_print("Simulating actions...\n");
+		}
+		else {
+			sprintf(cmd, "rm -rf /ss/safestrap/%s", arg.c_str());
+			__system(cmd);
+		}
+		operation_end(0, simulate);
+		return 0;
+	}
 	return -1;
 }
